Title: 가장 짧은 부분합, Time: 254ms, Memory: 17MB, Status: Wrong Answer - Codetree

+++ 250513/가장 짧은 부분합/shortest-subtotal.java
@@ -19,7 +19,7 @@ public class Main {
             answer[i] = Integer.parseInt(st.nextToken());
 
         int ans = MAX_NUM;
-        int sumVal = 0;
+        long sumVal = 0;
 
         int j = 1;
